feat: Repository 인터페이스 추가 및 적용
