+++ app/src/main/java/woowacourse/shopping/ui/MainActivity.kt
@@ -27,7 +27,6 @@ class MainActivity : AppCompatActivity() {
         setupView()
     }
 
-
     override fun onCreateOptionsMenu(menu: Menu?): Boolean {
         menuInflater.inflate(R.menu.cart_menu, menu)
         menu?.findItem(R.id.cart)?.actionView?.let { view ->
@@ -58,7 +57,7 @@ class MainActivity : AppCompatActivity() {
         viewModel.products.observe(this) {
             val adapter = ProductAdapter(
                 items = it,
-                onClickProduct = viewModel::addCartProduct
+                onClickProduct = viewModel::addCartProduct,
             )
             binding.rvProducts.adapter = adapter
         }
